+++ ReleaseNotes.md
@@ -2,7 +2,7 @@
 
 ## TODO
 
-GenericServices doesn't call access methods in classes which the top class inherits. e.g. in the example code below GenericServices would NOT call the method `InnerMethod`. This is different to properties, which reflection will find.
+1. GenericServices doesn't call access methods in classes which the top class inherits. e.g. in the example code below GenericServices would NOT call the method `InnerMethod`. This is different to properties, which reflection will find.
 
 ```c#
 class Outer : Inner
@@ -29,6 +29,11 @@ This code could be added to the `DecodedEntityClass` class to correct this, but 
 		inherited = inherited.BaseType;
 	}
 ```
+
+2. Add check for EntityStyle.HasNoKey and throw an exception when looking for a primary key in the DTO.
+3. Add a check for `QueryTrackingBehavior.NoTracking` on update and throw an exception - see issue #44.
+4. Have a look at issue #41. Seems that if all the DTO properties are marked as readonly then it fails.
+5. Update the exception handler in `SaveChangesExtensions` to follow the pattern in the EfCore.GenericEventRunner library.
 
 ## 3.1.0